kernel-tests: ion: Add auto-detection of heaps
Add auto-detection of heaps to make the unit test work on
all targets. The auto-detection of heaps works only on targets
with device tree.

Instead of hard-coding which heaps should be used for each test the
type of heap needed for the test is specified. During initialization
all the heaps are queried and the requested heap is assigned to the
test case. If a usable heap is not found the test will be
automatically skipped.

CRs-fixed: 479682
Change-Id: I98eaf57c0c2b4553400fe3ef6973c41a37f48418