bam_dmux_loopback: Add unit test module for bam_dmux
Bam dmux is often asked for various metrics, such as throughput and
latency. However, it is currently hard to isolate bam dmux and as such
it is hard to provide reasonable measurements for the said metrics of bam
dmux in isolation.

The loopback DLKM provides a means to isolate bam dmux and provides a way
to measure said metrics solely through the bam dmux portion of the data
path.

Change-Id: I9d48958fa0465038ec0df71933b22cf309b88c19