kernel-tests: Validate partial CPU context dumping from TZ and SDI
The CPU register dump format used by TZ and SDI for version 3 introduces
dumping of CPU context from SDI for those CPUs for which TZ was unable
to collect it. For instance, if TZ was able to collect CPU context for
CPUs 0, 2 and 3 but not for CPU 1, SDI collects it. Update the test for
validating this feature.

Change-Id: I7a1e3f16f80d9805bbb8465e297e598a706c4e14