memory_prof: fix segfault when alloc profile is bogus
Currently we return NULL (or an uninitialized pointer) when we fail to
parse the value of the --eval option or if the entire allocation profile
passed in with `-i' is bogus. In both cases, we proceed to use that
pointer. Since we probably don't have anything useful to do if we fail
to parse --eval or the allocation profile, just bail.

Change-Id: I4d6731535f9f602c0951a8d13695b00e2f7a2ffa