ipc_logging_test: Correct calculations in extraction_dump test case
In the extraction_dump test case, the calculation for how many messages
can go in each log page is incorrect, which causes the test to behave
incorrectly. In particular, partial pages do not start at "line 0" as
they should, and in multipage tests where the log wraps around, unknown
message types may be encountered.

Use 180 messages per page in the unit tests. This is an approximation
that avoids the issues described above.

Change-Id: I4d18a26da06d6243b8fbcd64993a5a0fc87afafc