ipc_logging_test: Add unit test for destructive reads
There is no unit test for the case that a full multi-page
log is being emptied (with a destructive read) before the
RAM dump is extracted. v0 cannot handle this case, while
v1 can. A test is needed to ensure both versions behave
as expected.

Add a unit test for this case.

Change-Id: I7a76f89ae5cdaf694a408e8888b443ebde3a4920