kernel-tests: iommu: Add additional unmapping corner case test
Add an additional test that tests a corner case in the IOMMU unmapping
code. The test will map the last page at the end of the first 1GB
virtual address space and then it will try to unmap. This will exercise
the code that frees the page tables.

This change also involves refactoring out the code that creates a domain
and attaches to the IOMMU. We want to reset everything between tests so
this allows us to easily attach and detach between different tests.

Change-Id: If5d40d83e4a96ee1eb76bd39ab5bad83fdf5cb66