ipc_logging_test: Add NULL check after allocation
In the non-destructive read test, data is allocated,
and it is assumed that the allocation is successful.
There is no NULL check present after the allocation.

Use the UT_ASSERT_PTR macro to assert that the data
pointer is not NULL after the allocation. If it is NULL,
go to the failure case, like other unit test failures.

Change-Id: I694875962219f89f4e6c98a98f9e9a836131ea35