kernel-tests: ip_accelerator: Add test for RNDIS Deaggregation Exception
Device-side USB chipset does not detect the end of a received USB
transfer that ends with a USB packet whose length is the 1024 bytes
for that endpoint. For this reason, the host must append a one-byte
zero packet to a data transfer that otherwise would have a length
that is a multiple of the 1024 bytes of the receiving endpoint.
USB RNDIS device must tolerate the appended byte. The MessageLength
field of the final REMOTE_NDIS_PACKET_MSG does not include this
appended byte. IPA HW generates Deaggregation Exception interrupt to
handle this case.

Change-Id: I4bada084921583cda668eaca877ac7bd1ddeb696