net: wireless: bcmdhd reduced the MAX_CNTL_TX_TIMEOUT to 1
Currently the recovery mechanism is not triggered because Wi-Fi
Framework will turn on Wi-Fi again as soon as we get resumed on timeout for the
first time.
Currently the recovery mechanism is triggered if we get resumend on  timeout two times
continually.
So, in order to trigger recovery mechanism, we need to the count from 2 to 1.

bug : 17755061 Wi-Fi stuck in Off position

Change-Id: I86a950243c742e561c3ac8a8722bd9a760331410
Signed-off-by: Ecco park <eccopark@broadcom.com>