kernel-tests: ensure pet work is stalled for apps bark wdog test
Watchdog driver was modified to allow pet work to get scheduled
on any cpu but the apps bark test work runs on cpu0 and hence
disables preemption only on cpu0 resulting in test failure if the
pet gets scheduled on a cpu other than cpu0.

Change the apps bark test to hotplug off all cpus except cpu0 so
that pet work is stalled.

Change-Id: Idb8ae754118bd883280a579fee879c37cb0b61d0