kernel-tests: glink: fix smd transitional race condition
The rx-done completion is set before the actual glink_rx_done() call is
made which opens up a race condition when intent re-use is specified
where the test code tries to re-use the intent, but the glink_rx_done()
call has not been completed leading to a communication failure and
timeout.

Defer setting the completion until after the glink_rx_done() is made to
eliminate the race condition.

Change-Id: Ic3e2804e62c26d7227e23c7d36a9f580cab13caf