kernel-tests: glink: Add unit test for LK/RPM handoff case
The LK bootloader communicates with the RPM and before it boots Linux,
it needs to tear down the link between LK and RPM so that Linux can
renegotiate the link with RPM when it starts booting.

Add unit test that simulates the teardown of the transport between the
local and remote system which simulates the LK to Linux handoff case.

Change-Id: Ia9e987f68f78f6c782a0b69b50d347045c2c30f9