ip_accelerator: add generic configuration test framework
Add the ability to configure IPA pipes from the user-space test
application.

Before this change, the configuration information was saved in
the IPA kernel test-module and the user-space test application sent an
index of that configuration.

With this new framework the user-space test application asks for a
specific configuration with all the required endpoint parameters.

This allows the test logic to be kept in the user-space test application.
The communication interface is based on a new IOCTL that is exposed from
the IPA test-module to its user-space test component.

Change-Id: I6b68d84be97797f1d0df33f8ea1ef1ee458102cd