kernel-tests: glink: Add SSR bypass test
Currently, the only way to test SSR remotely is to actually trigger
SSR. This is problematic if there is an issue with SSR handling in
other software - the system could fail before any G-Link SSR code
has a chance to run.

Add a test which directly calls the G-Link SSR handler. This will
perform all SSR functionality in the absence of an actual SSR event.
The test is prefixed with ft as this could put the system in a bad
state.

Change-Id: I5261fccc20dd8418856a4298ec2c131b72112e76