kernel-tests: glink: Add do_mock_negotiation_init() function
The do_mock_negotiation() function is currently used both in the unit
tests and in the DLKM's init function. If any of the asserts fail when
do_mock_negotiation is called from the init function, a NULL pointer
dereference will occur because NULL is passed in as the seq_file
pointer. However, performing mock negotiation is necessary as part of
the DLKM's initialization as it calls link_up for the mock transport.

Add a new function, do_mock_negotiation_init(), which performs the
negotiation and link_up and logs errors if they occur, but does not
take a seq_file pointer as an argument and does not perform any asserts.
Call this function in the DLKM's init function instead of
do_mock_negotiation().

Change-Id: Ic591efd33e776163efed6a0472d8e2b271ca0423