memory_prof: KPI and stability tests
Defines KPIs for concurrency and latency to:
1) Evaluate new features
2) Identify the effect of any memory related change on concurrency and latency
3) Find regressions, by comparing KPIs of different builds

In user terms, concurrency is the ability to sustain multiple apps in memory.
This is usually measured for e.g. like number of browser tabs that can be
sustained in background while we play a memory hungry game. Running a memory
hungry game translates to, running a memory hungry application in foreground.
So to simulate this, concurrency test forks tasks sequentially, and each forked
task allocates a certain amount of memory. Forking is done till a task is
killed. So concurrency index is calculated as the total memory that could be
allocated by all tasks till the first task was killed. The amount of memory each
task allocates is calculated using a mathematical progression. Concurrency test
uses geometric and arithmetic progressions. Geometric progression simulates
higher rates of allocation, while arithmetic progression simulates gradually
increasing rate of allocation. These tests are performed multiple times with
varying environments. For e.g. The newly forked tasks allocates memory while
being in foreground with adj 0, and are then moved to background. So there are
2 variants where the tasks are either moved to background or cached apps adj
levels. This is done to simulate situations to test features like adaptive LMK.
Also, some of the tests are performed with a parallel background write. This is
to include the effects of page cache on concurrency.

Concurrency test can be invoked using the following command:

	memory_prof -i [path to conc profile]

Latency in user terms is defined as the time taken to launch an application.
As far a memory is concerned this roughly translates to the time taken to
allocate memory by the newly launched application. Latency test measures the
latency by forking tasks till only one task remains. The time taken to allocate
memory is measured and the latency index is calculated as the time taken to
allocate 1MB of memory.

Latency test can be invoked using the following command:

	memory_prof -i [path to latency profile]

Also stability tests are intoduced, which is a set of tests run in a loop, which
can be run standalone or along with a parallel monkey or stress test to figure
out any kind of memory stability issues. At present it runs the latency test
with parallel file writes in a loop. This can exercise features like PPR,
adaptive LMK etc. In future when we come up with more features, tests to
exercise those can be added to this test.

To execute the stability test:

	memory_prof -i [path to stability profile]

Change-Id: I42b917f5018ae2687dc792f7550edfcc87c4282a