kernel-tests: glink: Change default tracer packet size to 256
The current default tracer packet size is 128, which is not large enough
to hold client RX and TX events for each direction in addition to the core
tracer packet events.

Increase the size to 256 so the tracer packet is large enough to
hold basic client RX and TX events by default.

Change-Id: Ic508e276e59e34b505a3bc7ec1197dfe3ad7e939
Signed-off-by: Steven Cahail <scahail@codeaurora.org>