kernel-tests: glink: Initialize open_cfg to NULL in mock ssr test
In the mock SSR test, the glink_open_config structure is not initialized
to anything. Functions exist in this structure which are called in the
core by default. The structure being uninitialized can result in
unexpected behavior, including corruption of G-Link channel handles.

Initialize the glink_open_config structure to NULL using memset().

Change-Id: I4eb20fc4a2774450e9eba2375dad338ad50bd949