kernel-tests: glink: Return number of bytes transmitted in loopback TX
The loopback transport's tx() function currently returns 0 on success.
The G-Link core now interprets the return value of a transport's TX
functions as the number of bytes written to the transport, and thus
considers a return value of 0 an error.

Modify the loopback transport's TX functions to return the number of
bytes written.

Change-Id: I4218ce112160bb16343502351ffd807a17c8bcfa