kernel-tests: glink: Return number of bytes transmitted in mock TX
The mock transport's tx() function currently returns 0 on success. The
G-Link core now interprets the return value of a transport's tx()
function as the number of bytes written to the transport, and thus
considers a return value of 0 an error.

Modify the mock transport's tx() function to return the number of bytes
written.

Change-Id: I64865442aec85d0bcc1194d61de5711a3fe05b80