ip_accelerator: test suspend IRQ not generated WA
This change tests suspend IRQ not generated HW limitation which
exists in ipa_v3 causing it not to send suspend IRQ upon aggregation
close if endpoint is suspended while aggregation is active.
This change adds a test that suspends an endpoint in the middle
of aggregation process and checks if the SW workaround is able to
overcome this limitation.

Change-Id: I0ddd883ffcf0add349e168069948bcf7092a032f