kernel-tests: modsign_test: Drop commands to restore original image
Modsign test does following verifications:

1. Verify loading/unloading of original signed module.
2. Update module image and verify that loading fails.
3. Update module image from step 2 to restore original image and
   verify that loading/unloading of module succeeds.

Step 2 and step 3 uses 'sed' command to insert/remove a character.
It has been found that behavior of those commands is inconsistent,
in the sense that, step 3 does not restore back original image.
This happens because, 'sed' command in step 2 inserts special
characters (in addition to the given character) to the image, which
are not removed by step 3.

Since step 1 and 2 are sufficient to verify signed module feature,
we drop sequence of commands for step 3.

Additionally, tainted image is cleaned up.

Change-Id: Ib00387f186ee6ad8cd534c8e76902c9dfbceeaa4