clearpad: Don't break when wakeup_gesture toggled during sleep
Let the wakeup_gesture attribute affect the next suspend of the touchscreen,
so that if wakeup_gesture is toggled while the touchscreen is suspended,
the driver/touchscreen doesn't get confused (because the driver wakes
it up assuming it is in normal suspend, but the touchscreen is in
tap-to-wake mode -- or vice versa).

This fixes the problem where phone calls cannot be answered with
tap-to-wake on.

Change-Id: I7fa5a826e84bd4a4b6cd25b8c249d2e83c628d3f
Signed-off-by: angelsl <hidingfromhidden@gmail.com>
(cherry picked from commit 6dc479f)