arm: DT: MSM8994: Move usbbam node, fix ssphy-qmp compatible string
Move the usbbam node to be a subnode of ssusb. This is a new
requirement after QC driver updates in kernel 3.18.
Moreover, change the compatible string of ssphy: the choice is
to either specify the offsets for it or change that string
to make the driver to choose the right data from structs
inside of it.