gpu: msm_legacy: A4xx: Check context when building preemption token
Adreno A4xx does not support draw context preemption and for
some (good!) reason it gets NULL to the function.
Check if context is NULL: if it is, then code wants to use
draw context to build preemption token, which is unsupported,
so build a generic preempt token based on generic preemption
register address on GPU.

If we have a context, then it will be a supported one: build
the preemption token as normal.